+++ src/main.rs
@@ -8,11 +8,12 @@ mod textures;
 mod shader;
 mod helpers;
 mod geometry;
+mod gl_program;
 
 use shader::*;
 use geometry::*;
 use textures::*;
-
+use gl_program::GLProgram;
 
 fn main() -> Result<(), anyhow::Error> {
     let mut glfw = glfw::init(glfw::FAIL_ON_ERRORS).unwrap();
@@ -80,6 +81,7 @@ fn main() -> Result<(), anyhow::Error> {
     triangle_1.scale(0.5);
     triangle_1.translade( 0.2, 0.2, 0.0);
     triangle_1.rotate(Angle::Z(45.0));
+    triangle_1.rotate(Angle::Y(45.0));
 
     triangle_2.scale(4.0);
